Consolidate exec implementations under utils/
