+++ way-cooler/utils/meson.build
@@ -1,4 +1,5 @@
 way_cooler_sources += files(
+	'exec.c',
 	'xkb_hash_set.c',
 )
